HM-3 patterns (PR fix)

+++ src/main/java/ru/astondevs/homework/arychev/patterns/adapter/PaymentProcessorImpl.java
@@ -1,12 +1,14 @@
 package ru.astondevs.homework.arychev.patterns.adapter;
 
 
+import java.math.BigDecimal;
+
 /**
  * Класс для обработки платежей в нашей системе
  */
 public class PaymentProcessorImpl implements PaymentProcessor {
     @Override
-    public void processPayment(String orderId, double amount) {
+    public void processPayment(String orderId, BigDecimal amount) {
         System.out.println("Processing payment for order " + orderId + " with amount $" + amount);
     }
 }
