+++ src/main/java/ru/astondevs/homework/arychev/patterns/adapter/ThirdPartyPayment.java
@@ -1,9 +1,11 @@
 package ru.astondevs.homework.arychev.patterns.adapter;
 
 
+import java.math.BigDecimal;
+
 /**
  * Интерфейс какой-то платежной системы
  */
 public interface ThirdPartyPayment {
-    void makePayment(String transactionId, double paymentAmount);
+    void makePayment(String transactionId, BigDecimal paymentAmount);
 }
